Revert debug
